Disabled spare(?) Timer script
Decided to hunt down the cause of the error that the 'game over canvas' wasn't assigned on the timer script, although everything was working as it should. Found that there's one timer script on the 'canvas' object, which has the game over canvas assigned, but there was also a second timer script on the ScoreManager object. I disabled this and the error is gone and nothing has broken after a few tests.